requirements_intel.txt
flask_socketio==5.3.4
pyqt5
requests
